requirements.txt
pyyaml~=6.0.1
beautifulsoup4~=4.12.3
requests~=2.31.0
bs4~=0.0.1
pytest~=7.4.0